// src/enums/RequestBotStatus.ts
export enum RequestBotStatus {
    WAITING = 0,
    ACCEPT = 1,
    ACCEPT_ERROR = 2,
    ACCEPT_ITEM_TO_GIVE_IS_NOT_NULL = 3,
    ACCEPT_TRADE_OFFER_NOT_FOUND = 4,
    ACCEPT_ERROR_CONFIRMATION = 5,
    ACCEPT_NOT_ALLOWED_STATE = 6,
    ACCEPT_UNSUPPORTED_CONFIRMATION = 6,
    ACCEPT_OK = 7,
    ACCEPT_ERROR_GET_RECEIVED_ITEMS = 8,
    SEND = 9,
    SEND_USER_TRADE_BAN = 10,
    SEND_USER_CANNOT_TRADE = 11,
    SEND_ERROR = 12,
    SEND_ERROR_CONFIRMATION = 13,
    SEND_OK = 14,
    SEND_ERROR_ITEM_TO_SEND_NULL = 15,
    SEND_COUNT_ADD_AND_ITEM_SET_LEN_IS_NOT_EQUALS = 16,
    SEND_ERROR_ON_CREATE_OFFER = 18,
    SEND_USER_CANNOT_GET_ITEMS = 19,
    COMPLETE = 20,
    EXPIRED = 21,
    COUNTERED = 22,
    DECLINED = 23,
    INVALID_ITEMS = 24,
    CANCELED_BY_SECOND_FACTOR = 25,
    CANCELED = 26,
    INVALID = 27
}